Updates to March 1st (566469f)

+++ library/src/main/java/org/apache/pdfbox/pdfparser/BaseParser.java
@@ -32,8 +32,7 @@ import android.util.Log;
  * This class is used to contain parsing logic that will be used by both the
  * PDFParser and the COSStreamParser.
  *
- * @author <a href="mailto:ben@benlitchfield.com">Ben Litchfield</a>
- * @version $Revision$
+ * @author Ben Litchfield
  */
 public abstract class BaseParser implements Closeable
 {
@@ -227,7 +226,7 @@ public abstract class BaseParser implements Closeable
 			{
 				throw new IOException("expected number, actual=" + number + " at offset " + genOffset);
 			}
-			COSObjectKey key = new COSObjectKey(((COSInteger) number).intValue(),
+			COSObjectKey key = new COSObjectKey(((COSInteger) number).longValue(),
 					((COSInteger) generationNumber).intValue());
 			retval = document.getObjectFromPool(key);
 		}
@@ -972,7 +971,7 @@ public abstract class BaseParser implements Closeable
 					if (po.get(po.size()-1) instanceof COSInteger)
 					{
 						COSInteger number = (COSInteger)po.remove( po.size() -1 );
-						COSObjectKey key = new COSObjectKey(number.intValue(), genNumber.intValue());
+						COSObjectKey key = new COSObjectKey(number.longValue(), genNumber.intValue());
 						pbo = document.getObjectFromPool(key);
 					}
 					else
@@ -1638,9 +1637,9 @@ public abstract class BaseParser implements Closeable
 	 * @return the object number being read.
 	 * @throws IOException if an I/O error occurs
 	 */
-	protected long readObjectNumber() throws IOException
+	protected int readObjectNumber() throws IOException
 	{
-		long retval = readLong();
+		int retval = readInt();
 		if (retval < 0 || retval >= OBJECT_NUMBER_THRESHOLD)
 		{
 			throw new IOException("Object Number '" + retval + "' has more than 10 digits or is negative");
